Symbols tables, binary search trees
Setup and submission

For this assignment you will write two programs and solve three exercises in the textbook. For the two programs, create a package called assignment1 in Eclipse and place them there. For the exercises, write the answers in a file called assignment1.{txt,doc,docx} that is either a text document or a Word document.

When you are finished, zip the assignment1 package into a zip file called assignment1.zip and so that it includes the two .java files. Add the exercises document to the zip file and submit to the D2L drop box provided.

Compute a GPA

Write a program called ComputeGPA that creates a symbol table and fills it with the following key-value pairs:

Letter	Points
A+	4.33
A	4.00
A-	3.67
B+	3.33
B	3.00
B-	2.67
C+	2.33
C	2.00
C-	1.67
D	1.00
F	0.00
It then reads a sequence of letter grades from standard input and computes and prints the GPA. The grades should be separated from each other with whitespace.

Write the program so that StdIn is redirected to a text file containing the grades. Use the fromFile(String filename) method for this. If the version of StdIn you have does not contain the fromFile method, add the following to the file in src/stdlib/StdIn.java:

	/**
	 * Redirect to a file.  This is a hack to get programs to work easily in eclipse.
	 * (Added by James Riely 2012/01/12.)
	 */
	public static void fromFile(String filename) {
		try {
			StdIn.scanner = new Scanner(new BufferedInputStream(new FileInputStream (filename)), StdIn.charsetName);
		} catch (FileNotFoundException e) {
			throw new Error (e.getMessage ());
		}
		StdIn.scanner.useLocale(StdIn.usLocale);
	}

	/**
	 * Redirect to a string.  This is a hack to get programs to work easily in eclipse.
	 * (Added by James Riely 2012/01/12.)
	 */
	public static void fromString(String s) {
		StdIn.scanner = new Scanner(s);
		StdIn.scanner.useLocale(StdIn.usLocale);
	}
Play a simple song

Write a program called PlaySimpleSong that first reads in a file notes_frequencies.txt where each line is a pair of strings separated by whitespace. The first string is the name of a musical note and the second its sound frequency as found on a piano. For example, the note A4 is paired with the frequency 440Hz and the note C4 with the frequency 261.626Hz. It fills a symbol table with these pairs where the note name is the key and the frequency is the value.

Next the program will read in a song file, where each line contains a note name and a duration in seconds, separated by whitespace. A sample file is sample_simple_song.txt, which plays every C for half a second.

To process a text file where each line contains a fixed set of data fields, use the method readLine(), which returns a string, split the string into the individual fields, and convert the fields into numeric values, when necessary.

To play each note, use this method:

	public static void playTone(double frequency, double duration) {
		final int sliceCount = (int) (StdAudio.SAMPLE_RATE * duration);
		final double[] slices = new double[sliceCount+1];
		for (int i = 0; i <= sliceCount; i++) {
			slices[i] = Math.sin(2 * Math.PI * i * frequency / StdAudio.SAMPLE_RATE);
		}
		StdAudio.play(slices);
	}
You will also have to import stdlib.StdAudio.