hash tables + BST
Add an iterator to the 2-3 tree class

Modify the class A4TwoThreeTree to be iterable. Use the class SimplerBSTIterable as a model. Test your code with A4TestTwoThreeTreeIterable. Call your class A4TwoThreeTreeIterable.

To make a collection class iterable requires three steps:

Indicate that the class implements the Iterable interface. In the example class SimplerBSTIterable, this meant making the first line of the class:
public class SimplerBSTIterable<K extends Comparable<? super K>, V> implements Iterable<K> {
This line specifies that an iterator over a SimplerBSTIterable object will return references to keys.
Implementing the java.util.Iterable requires this class to define a method with the signature public java.util.Iterator<K> iterator(). That method should construct and return a new Iterator object, where the Iterator class is defined in the next step.
Define an Iterator class that itself implements java.util.Iterator. This class must contain a constructor and three methods:
public boolean hasNext()
public K next()
public void remove() (You can implement this by having it throw an UnsupportedOperationException.)
Implement a simple hash table class

Write a generic hash table class that stores keys (but no values) in an array. Resolve collisions using linear probing, like the code in the book's class LinearProbingHashST (and which can be found in the algs34 package) but, unlike that code, when a key is removed, your class will not rehash following keys. Instead, it will use a second array of Boolean values, where each value is initially true. The interpretation of this second array is that if an entry is true and the corresponding slot in the hash array is null, this slot has never had a key in it. If an entry in the Boolean array is false and the corresponding hash slot is null, that means this slot had a key but it has since been deleted. You will need this Boolean array to perform a correct search on the hash table because your code will have to distinguish between when a null means that the key being sought is not there or that it means the search should skip to the next entry.

Here is the API you are to implement:

A4HashTable(int length): This constructs an empty hash table with the given length.
public int size(): Returns the number of keys that have been inserted.
public boolean isEmpty(): Returns whether the hash table is empty, that is, whether it has any keys in it.
public boolean contains(K key): Returns whether the given key is in the hash table.
public void put(K key): Inserts the given key into the hash table. Does nothing if the key is already there.
public void delete(K key): Removes the key from the table. Does nothing if the key was not there.